Approving cook
A cook no longer inherits a user, but instead stores the user as an attribute. This is because a cook is created from a user and this is some bs that doesn't work with java (you can't create a new derived object from an existing object without copying, and copying doesn't work either because in the DB it will be considered 2 different records).